More work updating to enable uid-centric approach
Updated not-found exceptions in recipes and ingredients to allow for an ingredient or a recipe approach.
Fix a bug where UnknownNutrientFlagError wanted a str uid instead of an int.
Fix an issue where NutrientDTO was still using a string to reference parent name, instead of uid, and a similar issue in NutrientFlagDefDTO, where str was used to reference parent and also excluded nutrients.